requirements.txt
# Web framework and server
fastapi>=0.115.14
uvicorn[standard]>=0.35.0

# Database & cache layers
sqlalchemy>=2.0.41
psycopg2-binary>=2.9.10
redis>=6.2.0

# Security
bcrypt>=4.3.0
python-jose[cryptography]>=3.5.0
passlib[bcrypt]>=1.7.4
itsdangerous>=2.2.0

# Validation
pydantic>=2.11.7
pydantic-settings>=2.10.1
email-validator>=2.2.0

# Networking & I/O
httpx>=0.28.1
aiohttp>=3.12.13
aiofiles>=24.1.0
python-multipart>=0.0.20

# File parsing
python-docx>=1.2.0
pypdf2>=3.0.1
pdfplumber>=0.11.0
mammoth>=1.9.1
python-magic>=0.4.27

# OCR and image processing
pytesseract>=0.3.10
Pillow>=10.0.0
PyMuPDF>=1.26.0

# Scraping & text cleaning
beautifulsoup4>=4.13.4
requests>=2.32.4
trafilatura>=2.0.0

# LLMs & embeddings (Production - Bedrock only)
boto3>=1.39.3
nltk>=3.8.1

# Config & observability
python-dotenv>=1.1.1
structlog>=25.4.0
jinja2>=3.1.6

# Rate limiting and middleware
slowapi>=0.1.9

# Additional AI/ML utilities (Essential only)
numpy>=1.24.0

# Monitoring and health checks
prometheus-client>=0.20.0
psutil>=5.9.0

# Production server
gunicorn>=21.2.0

# Production server
gunicorn>=21.2.0
